+++ src/main/java/com/github/jlangch/venice/impl/docgen/cheatsheet/DocGenerator.java
@@ -1908,11 +1908,11 @@ public class DocGenerator {
 		util.addItem(getDocItem("ns-list"));
 		util.addItem(getDocItem("namespace"));
 
-//		final DocSection alias = new DocSection("Alias", "namespace.alias");
-//		all.addSection(alias);
-//		alias.addItem(getDocItem("ns-alias"));
-//		alias.addItem(getDocItem("ns-aliases"));
-//		alias.addItem(getDocItem("ns-unalias"));
+		final DocSection alias = new DocSection("Alias", "namespace.alias");
+		all.addSection(alias);
+		alias.addItem(getDocItem("ns-alias"));
+		alias.addItem(getDocItem("ns-aliases"));
+		alias.addItem(getDocItem("ns-unalias"));
 		
 		return section;
 	}